refactor: moving function location
